requirements.txt
pytest==6.2.1
requests==2.25.1
requests-mock==1.8.0
PyHamcrest==2.0.2
Faker==5.4.0
